+++ nbproject/private/private.properties
@@ -1,4 +1,5 @@
-deploy.ant.properties.file=C:\\Users\\ACER\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
-j2ee.server.home=C:/Program Files/Apache Software Foundation/Tomcat 10.1
-j2ee.server.instance=tomcat101:home=C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1
-user.properties.file=C:\\Users\\ACER\\AppData\\Roaming\\NetBeans\\17\\build.properties
+deploy.ant.properties.file=C:\\Users\\Admin\\AppData\\Roaming\\NetBeans\\22\\tomcat101.properties
+j2ee.server.home=C:/Users/Admin/Downloads/apache-tomcat-10.1.30/apache-tomcat-10.1.30
+j2ee.server.instance=tomcat101:home=C:\\Users\\Admin\\Downloads\\apache-tomcat-10.1.30\\apache-tomcat-10.1.30
+selected.browser=default
+user.properties.file=C:\\Users\\Admin\\AppData\\Roaming\\NetBeans\\22\\build.properties
