+++ nbproject/private/private.xml
@@ -3,8 +3,11 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/D:/ChildrenCare/src/java/controller/homePageServlet.java</file>
-            <file>file:/D:/ChildrenCare/web/WEB-INF/web.xml</file>
+            <file>file:/D:/swp/ChildrenCare/src/java/model/User.java</file>
+            <file>file:/D:/swp/ChildrenCare/web/index.html</file>
+            <file>file:/D:/swp/ChildrenCare/web/login.jsp</file>
+            <file>file:/D:/swp/ChildrenCare/web/WEB-INF/web.xml</file>
+            <file>file:/D:/swp/ChildrenCare/src/java/controller/LoginServlet.java</file>
         </group>
     </open-files>
 </project-private>
