+++ web/WEB-INF/web.xml
@@ -4,8 +4,20 @@
         <servlet-name>homePageServlet</servlet-name>
         <servlet-class>controller.homePageServlet</servlet-class>
     </servlet>
+    <servlet>
+        <servlet-name>homePageServlet</servlet-name>
+        <servlet-class>controller.homePageServlet</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>LoginServlet</servlet-name>
+        <servlet-class>controller.LoginServlet</servlet-class>
+    </servlet>
     <servlet-mapping>
         <servlet-name>homePageServlet</servlet-name>
         <url-pattern>/home</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>LoginServlet</servlet-name>
+        <url-pattern>/login</url-pattern>
     </servlet-mapping>
 </web-app>
